allsoc: est: hardcoded macport removed
- Hard coded MAC port replaced with
  sysconfig generated mac port for
  enet_cpsw_est example.

Fixes: SITSW-12285

+++ examples/enet_cpsw_est/V0/enet_cpsw_est_common.h
@@ -164,7 +164,7 @@ typedef struct EnetApp_Obj_s
     bool usingDfltSched;
 
     /* Number of MAC ports being used */
-    uint32_t macPortNum;
+    uint8_t macPortNum;
 
     /* This core's id */
     uint32_t coreId;
